create pygfx classes by modifying builtins
The existing pygfx classes for the wobject/material/shader are a good
jumping off point for our own code. We make a few key modifications to
the builtins:

1. Data is stored in the wobject. Previously, the wobject itself didn't
really hold much. Instead, the geometry of the wobject held the texture
with the data. The material held info about rendering. The design
decision made here is to store all data in the wobject and info about
how to render in the material. The geometry is calculated from the data
and created by the wobject itself.

2. The shader is split into multiple files. There is no longer a single
large shader for everything. Instead, a custom package is created for
the jinja templates to hold the shader code for our shader. Different
functions are split into files. Certain functions that are imported in
the original shader are also modified.

This commit also includes tests for setting up a single screen and
rendering to an offscreen canvas. The main script features a similar
version of the script, but rendering to rendercanvas.auto. Note that the
main script checks for a specific GPU on my personal workstation.