add benchmarking script for concurrent texture updates
A planned feature of this renderer is to update the ring buffer texture
during renders via a background thread. This would introduce a race
condition during renders which we would have to solve. However, the
WebGPU standard describes this as undefined behavior. This script runs a
benchmark for with/without performance of a background thread.