reorganize tests for volumes that reuse the same basic data
Most of the existing tests use the same data and volume sizes. In order
to reuse this, the fixtures for the data are extracted out into
conftest.py. Because not all tests will want to have their fixture
namespace polluted by this, the tests using these fixtures are moved
into a new directory.