extract out method for redrawing an offscreen scene in tests
Sometimes, tests might want to change things in the scene in-between
renders. `GfxContext.render_object` fully sets up the scene for each
call, which is not the desired behavior. 

To solve this, `GfxContext._canvas.draw`, which returns an array of the
pixels on screen, is pulled out into `GfxContext.draw` so that it can be
called multiple times to redraw the scene.