create a ring buffer implementation
This implementation essentially brute forces the requires chunks. It
diffs the chunks for the new and old camera positions and returns the
operations from that. Because the brute force method might be slow,
future implementations could implement smarter algorithims or be written
in a different language with FFI (e.g., Rust or C).

There is also a second implementation inheriting from the original that
knows about how chunks work and will also filter out negative
assignments. This implementation should be the one used in the actual
app, with the parent serving to hold all the logic.