replace existing ring buffer with a new implementation
The existing ring buffer implementation is very slow and does not match
the API we want. It is replaced with a new implementation that should be
prefered. The scripts will be updated in the future to make use of the
new implementation.

Note that even though the new implementation is called a "wrapping
buffer", many parameter names and in-code references refer to it as a
ring buffer still. This should be changed in the future.