add distance checks into the shader
We need some flag to hide data that isn't supposed to be rendered. Since
we won't load anything outside our view distance, we use that as a way
to determine when to stop sampling data for now. To make the view 
distance calculation easier, we also add in an extra uniform.

This commit breaks some tests. We remove
"test_ring_buffer_volume_wrapping", since the ring buffer wrapping is no
longer a behavior that is expected. The view distance (ring buffer n) is
also adjusted for other tests to ensure the entire object is rendered.