split tests out into a seperate pixi feature
We might want to run tests outside of a dev environment (e.g., if we run
them in a CI runner). Splitting this out allows us to do so. Note that
this only lays out the groundwork for this. We don't actually create a
new environment (yet) to install. 

We might also want to look into making ruff part of the test feature,
since in this CI example we'd probably also want to check ruff lint
rules along with out tests.