create a uniform for the wrapping buffer
The wrapping buffer needs to provide uniform data to the GPU. By
creating the uniform buffer on the wrapping buffer class itself, the
shader and the wobject won't need to know anything about the specifics
from the data passed.

Another thing that could be added would be merge the wrapping buffer's
uniform into the wobject's uniform, so the shader won't need to know
about the wrapping buffer's existence at all.