reintroduce depth picking to the shader
Although we removed depth picking earlier, we need to have it enabled so
we can add other pygfx wobjects (segmentation meshes) into the scene.
This commit also includes a more detailed explanation on what exactly
the depth picking algorithm is doing, since the previous version was
very vague with the explanation.