add user defined color mapping for labels
Users will want to provide their own colormaps for labels. This commit
introduces a new buffer on the material that lets users provide HSV
values for colors.

This implementation uses a mod on the label ID to select a color, but in
the future users might want to provide their own map. Furthermore, this
implementation only lets users select hue and saturation values. The V
channel is determined by the sampled intensity.