force cast writes to textures in WrappingBuffer
If our input data is of different types, we need to update the types of
our shader code to match. We instead fix the types we write to the
textures to be float32 for intensity data and uint32 for segmentation 
data. This probably isn't the best approach since we might have uint32
input data.

Other fixes might be to just template the types of everything, perform
casts in the shader (i.e., use overloads for textureLoad but then just
immediately cast it), or use some mystery overloading/generic technique
that I just don't know about yet.