add comments on design choices for the shader
The shader class isn't the most straightforward since it's collection of
code snippets found from different shader classes in the pygfx gallery
and source code. 

1. We use `register_wgpu_render_function` as a decorator to implicitly
call it with one argument. On the pygfx side, this creates a list of
shader classes with our shader class as the only element. If we wanted
to register shaders sequentially, we need to call the function at the
top level with multiple arguments which should be run in order.

2. We skip a couple levels of the inheritance hierarchy and calling
`super().__init__()` in our shader class. Because inheritance is used
here as a way to stay DRY, we need to skip some of the logic in the base
classes that we don't want.